Add hardened GnuPG configuration file
This configuration has some safer settings than the default configuration, and includes a keyserver pool instead of a single keyserver for more reliable transfer.
It was weakened somewhat from the Riseup OpenPGP version, in that the communication with the keyserver is over HTTP instead of HTTPS (hkp instead of hkps), for higher usability in the workshop.